Summary: Fix inconsistency in trip count cache and trip count type causing possible truncation of values stored in trip count cache. Reviewed by: Chawla, Pankaj Closed CQ(s): Compiler self build on Windows Rurl: https://ir-codecollab.intel.com/ui#review:id=43672 Old/split llvm commit ID: 58b80109c09294020a83da1e89ba420c1fa8583f

+++ llvm/lib/Transforms/Intel_LoopTransforms/HIRCompleteUnroll.cpp
@@ -122,7 +122,7 @@ private:
   SmallVector<HLLoop *, 32> CandidateLoops;
 
   // Level to TripCount mapping. This is just used during the analysis phase.
-  unsigned TripCountCache[MaxLoopNestLevel + 1];
+  int64_t TripCountCache[MaxLoopNestLevel + 1];
 
   /// \brief Computes the trip count necessary for complete unrolling.
   /// This routine handles cases such as triangular loops.
